Begin replacing compatibility mode Note: opening this in intelij will show a bunch of errors
