+++ src/main/java/net/countercraft/movecraft/api/MovecraftLocation.java
@@ -15,7 +15,7 @@
  *     along with Movecraft.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.countercraft.movecraft.utils;
+package net.countercraft.movecraft.api;
 
 import org.bukkit.Location;
 import org.bukkit.World;
