+++ src/main/java/net/countercraft/movecraft/api/WorldHandler.java
@@ -1,8 +1,6 @@
 package net.countercraft.movecraft.api;
 
-import net.countercraft.movecraft.craft.Craft;
-import net.countercraft.movecraft.utils.MovecraftLocation;
-import net.countercraft.movecraft.utils.Rotation;
+import net.countercraft.movecraft.api.craft.Craft;
 import org.jetbrains.annotations.NotNull;
 
 public abstract class WorldHandler {
