+++ src/main/java/net/countercraft/movecraft/api/craft/Craft.java
@@ -15,15 +15,10 @@
  *     along with Movecraft.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.countercraft.movecraft.craft;
+package net.countercraft.movecraft.api.craft;
 
-import net.countercraft.movecraft.async.AsyncManager;
-import net.countercraft.movecraft.async.detection.DetectionTask;
-import net.countercraft.movecraft.async.rotation.RotationTask;
-import net.countercraft.movecraft.async.translation.TranslationTask;
-import net.countercraft.movecraft.async.translation.TranslationTaskData;
-import net.countercraft.movecraft.utils.MovecraftLocation;
-import net.countercraft.movecraft.utils.Rotation;
+import net.countercraft.movecraft.api.Rotation;
+import net.countercraft.movecraft.api.MovecraftLocation;
 import org.bukkit.ChatColor;
 import org.bukkit.World;
 import org.bukkit.block.Sign;
@@ -32,13 +27,20 @@ import org.bukkit.entity.Player;
 import java.util.HashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-public class Craft {
-    private final CraftType type;
-    private int[][][] hitBox;
-    private MovecraftLocation[] blockList;
-    private World w;
+/*import net.countercraft.movecraft.async.AsyncManager;
+import net.countercraft.movecraft.async.rotation.RotationTask;
+import net.countercraft.movecraft.async.translation.TranslationTask;
+import net.countercraft.movecraft.async.translation.TranslationTaskData;*/
+
+public abstract class Craft {
+    protected final CraftType type;
+    protected int[][][] hitBox;
+    protected MovecraftLocation[] blockList;
+    protected World w;
     private AtomicBoolean processing = new AtomicBoolean();
-    private int minX, minZ, maxHeightLimit;
+    protected int minX;
+    protected int minZ;
+    private int maxHeightLimit;
     private boolean cruising;
     private boolean sinking;
     private boolean disabled;
@@ -121,65 +123,9 @@ public class Craft {
         this.hitBox = hitBox;
     }
 
-    public void detect(Player player, Player notificationPlayer, MovecraftLocation startPoint) {
-        AsyncManager.getInstance().submitTask(new DetectionTask(this, startPoint, type.getMinSize(), type.getMaxSize(), type.getAllowedBlocks(), type.getForbiddenBlocks(), type.getForbiddenSignStrings(), player, notificationPlayer, w), this);
-    }
+    public abstract void detect(Player player, Player notificationPlayer, MovecraftLocation startPoint);
 
-    public void translate(int dx, int dy, int dz) {
-        // check to see if the craft is trying to move in a direction not permitted by the type
-        if (!this.getType().allowHorizontalMovement() && !this.getSinking()) {
-            dx = 0;
-            dz = 0;
-        }
-        if (!this.getType().allowVerticalMovement() && !this.getSinking()) {
-            dy = 0;
-        }
-        if (dx == 0 && dy == 0 && dz == 0) {
-            return;
-        }
-
-        if (!this.getType().allowVerticalTakeoffAndLanding() && dy != 0 && !this.getSinking()) {
-            if (dx == 0 && dz == 0) {
-                return;
-            }
-        }
-
-        // find region that will need to be loaded to translate this craft
-        int cminX = minX;
-        int cmaxX = minX;
-        if (dx < 0)
-            cminX = cminX + dx;
-        int cminZ = minZ;
-        int cmaxZ = minZ;
-        if (dz < 0)
-            cminZ = cminZ + dz;
-        for (MovecraftLocation m : blockList) {
-            if (m.getX() > cmaxX)
-                cmaxX = m.getX();
-            if (m.getZ() > cmaxZ)
-                cmaxZ = m.getZ();
-        }
-        if (dx > 0)
-            cmaxX = cmaxX + dx;
-        if (dz > 0)
-            cmaxZ = cmaxZ + dz;
-        cminX = cminX >> 4;
-        cminZ = cminZ >> 4;
-        cmaxX = cmaxX >> 4;
-        cmaxZ = cmaxZ >> 4;
-
-
-        // load all chunks that will be needed to translate this craft
-        for (int posX = cminX - 1; posX <= cmaxX + 1; posX++) {
-            for (int posZ = cminZ - 1; posZ <= cmaxZ + 1; posZ++) {
-                if (!this.getW().isChunkLoaded(posX, posZ)) {
-                    this.getW().loadChunk(posX, posZ);
-                }
-            }
-        }
-
-        AsyncManager.getInstance().submitTask(new TranslationTask(this, new TranslationTaskData(dx, dz, dy, getBlockList(), getHitBox(), minZ, minX, type.getMaxHeightLimit(), type.getMinHeightLimit())), this);
-    }
+    public abstract void translate(int dx, int dy, int dz);
 
     public void resetSigns(boolean resetCruise, boolean resetAscend, boolean resetDescend) {
         for (MovecraftLocation aBlockList : blockList) {
@@ -205,49 +151,9 @@ public class Craft {
         }
     }
 
-    public void rotate(Rotation rotation, MovecraftLocation originPoint) {
-        // find region that will need to be loaded to rotate this craft
-        int cminX = minX;
-        int cmaxX = minX;
-        int cminZ = minZ;
-        int cmaxZ = minZ;
-        for (MovecraftLocation m : blockList) {
-            if (m.getX() > cmaxX)
-                cmaxX = m.getX();
-            if (m.getZ() > cmaxZ)
-                cmaxZ = m.getZ();
-        }
-        int distX = cmaxX - cminX;
-        int distZ = cmaxZ - cminZ;
-        if (distX > distZ) {
-            cminZ -= (distX - distZ) / 2;
-            cmaxZ += (distX - distZ) / 2;
-        }
-        if (distZ > distX) {
-            cminX -= (distZ - distX) / 2;
-            cmaxX += (distZ - distX) / 2;
-        }
-        cminX = cminX >> 4;
-        cminZ = cminZ >> 4;
-        cmaxX = cmaxX >> 4;
-        cmaxZ = cmaxZ >> 4;
+    public abstract void rotate(Rotation rotation, MovecraftLocation originPoint);
 
-
-        // load all chunks that will be needed to rotate this craft
-        for (int posX = cminX; posX <= cmaxX; posX++) {
-            for (int posZ = cminZ; posZ <= cmaxZ; posZ++) {
-                if (!this.getW().isChunkLoaded(posX, posZ)) {
-                    this.getW().loadChunk(posX, posZ);
-                }
-            }
-        }
-
-        AsyncManager.getInstance().submitTask(new RotationTask(this, originPoint, this.getBlockList(), rotation, this.getW()), this);
-    }
-
-    public void rotate(Rotation rotation, MovecraftLocation originPoint, boolean isSubCraft) {
-        AsyncManager.getInstance().submitTask(new RotationTask(this, originPoint, this.getBlockList(), rotation, this.getW(), isSubCraft), this);
-    }
+    public abstract void rotate(Rotation rotation, MovecraftLocation originPoint, boolean isSubCraft);
 
     public int getMaxX() {
         return minX + hitBox.length;
