+++ src/main/java/net/countercraft/movecraft/commands/ManOverboardCommand.java
@@ -1,7 +1,7 @@
 package net.countercraft.movecraft.commands;
 
 import net.countercraft.movecraft.config.Settings;
-import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.api.craft.Craft;
 import net.countercraft.movecraft.craft.CraftManager;
 import net.countercraft.movecraft.localisation.I18nSupport;
 import org.bukkit.Location;
