+++ src/main/java/net/countercraft/movecraft/commands/PilotCommand.java
@@ -1,11 +1,12 @@
 package net.countercraft.movecraft.commands;
 
-import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.api.MovecraftLocation;
+import net.countercraft.movecraft.api.craft.Craft;
+import net.countercraft.movecraft.api.craft.CraftType;
 import net.countercraft.movecraft.craft.CraftManager;
-import net.countercraft.movecraft.craft.CraftType;
+import net.countercraft.movecraft.craft.ICraft;
 import net.countercraft.movecraft.localisation.I18nSupport;
 import net.countercraft.movecraft.utils.MathUtils;
-import net.countercraft.movecraft.utils.MovecraftLocation;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabExecutor;
@@ -43,7 +44,7 @@ public class PilotCommand implements TabExecutor {
             if (oldCraft != null) {
                 CraftManager.getInstance().removeCraft(oldCraft);
             }
-            Craft newCraft = new Craft(craftType, player.getWorld());
+            Craft newCraft = new ICraft(craftType, player.getWorld());
             MovecraftLocation startPoint = MathUtils.bukkit2MovecraftLoc(player.getLocation());
             newCraft.detect(player, player, startPoint);
         } else {
