+++ src/main/java/net/countercraft/movecraft/commands/ReleaseCommand.java
@@ -1,6 +1,6 @@
 package net.countercraft.movecraft.commands;
 
-import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.api.craft.Craft;
 import net.countercraft.movecraft.craft.CraftManager;
 import net.countercraft.movecraft.localisation.I18nSupport;
 import org.bukkit.Bukkit;
