+++ src/main/java/net/countercraft/movecraft/compat/v1_12_R1/ICraftTranslateCommand.java
@@ -1,0 +1,4 @@
+package net.countercraft.movecraft.compat.v1_12_R1;
+
+public class ICraftTranslateCommand {
+}
