+++ src/main/java/net/countercraft/movecraft/utils/MathUtils.java
@@ -17,7 +17,9 @@
 
 package net.countercraft.movecraft.utils;
 
-import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.api.MovecraftLocation;
+import net.countercraft.movecraft.api.Rotation;
+import net.countercraft.movecraft.api.craft.Craft;
 import org.bukkit.Location;
 
 public class MathUtils {
